+++ routes/servers/endpoints/patch_server_settings/route.go
@@ -35,6 +35,7 @@ func updateServerArgs(server types.ServerSettingsUpdate) []any {
 		server.NSFW,
 		server.CaptchaOptOut,
 		server.LoginRequiredForInvite,
+		server.ShowEmojis,
 	}
 }